requirements/common.txt
Django==2.2.4
Pillow==6.2.1
django-crispy-forms==1.7.2
djangorestframework==3.11.0
python-dateutil==2.8.0
sorl-thumbnail==12.6.3
